Updated ToggleItem
