--- UiToggleItem/UiToggleItem.module.less
@@ -1,9 +1,0 @@
-.toggleItem {
-	/* Placeholder for CSS overrides */
-	.slot,
-	.after,
-	.before,
-	.hidden {
-		/* Public Class Names */
-	}
-}
